+++ src/Areas/VirtualDesktop/Commands/VirtualDesktopJsonContext.cs
@@ -10,9 +10,12 @@ namespace AzureMcp.Areas.VirtualDesktop.Commands;
 
 [JsonSerializable(typeof(HostpoolListCommand.HostPoolListCommandResult))]
 [JsonSerializable(typeof(SessionHostListCommand.SessionHostListCommandResult))]
+[JsonSerializable(typeof(SessionHostUserSessionListCommand.SessionHostUserSessionListCommandResult))]
 [JsonSerializable(typeof(HostPool))]
 [JsonSerializable(typeof(Models.SessionHost))]
+[JsonSerializable(typeof(UserSession))]
 [JsonSerializable(typeof(List<Models.SessionHost>))]
+[JsonSerializable(typeof(List<UserSession>))]
 [JsonSourceGenerationOptions(PropertyNamingPolicy = JsonKnownNamingPolicy.CamelCase, DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull)]
 internal sealed partial class VirtualDesktopJsonContext : JsonSerializerContext
 {